+++ otj-server-reactive/src/main/java/com/opentable/server/reactive/ReactiveServerCommonConfiguration.java
@@ -22,6 +22,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
+import com.opentable.conservedheaders.reactive.ReactiveServerConservedHeadersConfiguration;
 import com.opentable.metrics.reactive.HealthHttpReactiveConfiguration;
 import com.opentable.metrics.reactive.MetricsHttpReactiveConfiguration;
 import com.opentable.server.EmbeddedJettyConfiguration;
@@ -40,7 +41,10 @@ import com.opentable.server.reactive.webfilter.BackendInfoWebFilterConfiguration
         EmbeddedJettyConfiguration.class,
         EmbeddedReactiveJetty.class,
         // Log server request information
+        // TODO: Remove this after updating servers that use ServerRequestLog directly
         ReactiveServerLoggingConfiguration.class,
+        // Conserved headers
+        ReactiveServerConservedHeadersConfiguration.class,
         // Default health check
         HealthHttpReactiveConfiguration.class,
         // Metrics for http